requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.22
openai>=1.12.0
pydantic>=2.5.2
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.6
tiktoken>=0.5.2
python-dateutil>=2.8.2
PyNaCl>=1.5.0  # For voice support
sounddevice>=0.4.6  # For audio recording
soundfile>=0.12.1  # For audio file handling
numpy>=1.24.0  # Required for audio processing